+++ scripts/ci_validate_modification.ts
@@ -1,15 +1,18 @@
 import * as path from 'path';
 import { promises as fs } from 'fs';
 import * as cp from 'child_process';
-import { DataJson, BenchmarkSuites, SCRIPT_PREFIX } from '../src/write';
+import { BenchmarkSuites, DataJson, SCRIPT_PREFIX } from '../src/write';
 import { VALID_TOOLS } from '../src/config';
 import { Benchmark } from '../src/extract';
-import { diff, Diff, DiffNew, DiffEdit, DiffArray } from 'deep-diff';
-import deepEq = require('deep-equal');
+import { diff, Diff, DiffArray, DiffEdit, DiffNew } from 'deep-diff';
 import { getServerUrl } from '../src/git';
+import assert from 'assert';
+import deepEq = require('deep-equal');
 
 function help(): never {
-    throw new Error('Usage: node ci_validate_modification.js before_data.js "bechmark name"');
+    throw new Error(
+        'Usage: node ci_validate_modification.js before_data.js "benchmark name" [benchmark-data-repository-directory]',
+    );
 }
 
 async function exec(cmd: string): Promise<string> {
@@ -47,7 +50,7 @@ function validateDataJson(data: DataJson) {
     for (const benchName of Object.keys(suites)) {
         for (const suite of suites[benchName]) {
             const { commit, tool, date, benches } = suite;
-            if (!(VALID_TOOLS as string[]).includes(tool)) {
+            if (!(VALID_TOOLS as ReadonlyArray<string>).includes(tool)) {
                 throw new Error(`Invalid tool ${tool}`);
             }
             if (!commitUrlMatcher.test(commit.url) && !/\/pull\/\d+\/commits\/[a-f0-9]+$/.test(commit.url)) {
@@ -184,10 +187,27 @@ function validateBenchmarkResultMod<T>(diff: Diff<T>, expectedBenchName: string,
     }
 }
 
+function validateDiff(beforeJson: DataJson, afterJson: DataJson, expectedBenchName: string) {
+    const diffs = diff(beforeJson, afterJson);
+    console.log('Validating diffs:', diffs);
+
+    if (!diffs || diffs.length !== 2) {
+        throw new Error('Number of diffs are incorrect. Exact 2 diffs are expected');
+    }
+
+    console.log('Validating lastUpdate modification');
+    validateLastUpdateMod(diffs[0]);
+
+    console.log('Validating benchmark result modification');
+    validateBenchmarkResultMod(diffs[1], expectedBenchName, afterJson.entries);
+
+    console.log('👌');
+}
+
 async function main() {
     console.log('Start validating modifications by action with args', process.argv);
 
-    if (process.argv.length !== 4) {
+    if (process.argv.length !== 4 && process.argv.length !== 5) {
         help();
     }
 
@@ -203,6 +223,11 @@ async function main() {
 
     const beforeDataJs = path.resolve(process.argv[2]);
     const expectedBenchName = process.argv[3];
+    const benchmarkDataDirectory = process.argv[4];
+
+    const additionalGitParams = benchmarkDataDirectory
+        ? `--work-tree=${benchmarkDataDirectory} --git-dir=${benchmarkDataDirectory}/.git`
+        : '';
 
     console.log('Validating modifications by action');
     console.log(`  data.js before action: ${beforeDataJs}`);
@@ -211,14 +236,14 @@ async function main() {
     const beforeJson = await readDataJson(beforeDataJs);
 
     console.log('Validating current branch');
-    const branch = await exec('git rev-parse --abbrev-ref HEAD');
+    const branch = await exec(`git ${additionalGitParams} rev-parse --abbrev-ref HEAD`);
     if (branch === 'gh-pages') {
         throw new Error(`Current branch is still on '${branch}'`);
     }
 
     console.log('Retrieving data.js after action');
-    await exec('git checkout gh-pages');
-    const latestCommitLog = await exec('git log -n 1');
+    await exec(`git ${additionalGitParams} checkout gh-pages`);
+    const latestCommitLog = await exec(`git ${additionalGitParams} log -n 1`);
 
     console.log('Validating auto commit');
     const commitLogLines = latestCommitLog.split('\n');
@@ -239,27 +264,25 @@ async function main() {
         throw new Error(`Unexpected auto commit message in log '${latestCommitLog}'`);
     }
 
-    const afterJson = await readDataJson('dev/bench/data.js');
-    await exec('git checkout -');
+    const dataResults = await Promise.allSettled([
+        readDataJson('benchmark-data-repository/dev/bench/data.js'),
+        readDataJson('dev/bench/data.js'),
+    ]);
+
+    const jsonResults = dataResults
+        .filter((res): res is PromiseFulfilledResult<DataJson> => res.status === 'fulfilled')
+        .map((res) => res.value);
+
+    assert(jsonResults.length > 0 && jsonResults.length <= 2, 'Maximum 2 data.js files should be present in the repo');
+
+    const afterJson = jsonResults[0];
+    await exec(`git ${additionalGitParams} checkout -`);
 
     console.log('Validating data.js both before/after action');
     validateDataJson(beforeJson);
     validateDataJson(afterJson);
 
-    const diffs = diff(beforeJson, afterJson);
-    console.log('Validating diffs:', diffs);
-
-    if (!diffs || diffs.length !== 2) {
-        throw new Error('Number of diffs are incorrect. Exact 2 diffs are expected');
-    }
-
-    console.log('Validating lastUpdate modification');
-    validateLastUpdateMod(diffs[0]);
-
-    console.log('Validating benchmark result modification');
-    validateBenchmarkResultMod(diffs[1], expectedBenchName, afterJson.entries);
-
-    console.log('👌');
+    validateDiff(beforeJson, afterJson, expectedBenchName);
 }
 
 main().catch((err) => {