+++ test/data/write/benchmark-data-repository/data-dir/original_data.js
@@ -1,0 +1,23 @@
+window.BENCHMARK_DATA = {
+    "lastUpdate": 1574927128603,
+    "repoUrl": "https://github.com/user/repo",
+    "entries": {
+        "Test benchmark": [
+            {
+                "commit": {
+                    "author": { "email": "dummy@example.com", "name": "User", "username": "user" },
+                    "committer": { "email": "dummy@example.com", "name": "User", "username": "user" },
+                    "distinct": false,
+                    "id": "prev commit id",
+                    "message": "dummy message",
+                    "timestamp": "dummy stamp",
+                    "tree_id": "dummy tree id",
+                    "url": "https://github.com/user/repo/commit/prev commit id"
+                },
+                "date": 1574927127603,
+                "tool": "cargo",
+                "benches": [{ "name": "bench_fib_10", "range": "± 20", "unit": "ns/iter", "value": 100 }]
+            }
+        ]
+    }
+}